Fix Print Function in USART and LCD

+++ ECLIPSE_PROJECTS/AVR_DRIVERS/inc/Lcd.h
@@ -41,7 +41,7 @@ typedef struct
 void Lcd_Init(void);
 void Lcd_DisplayCharacter (u8 character);
 void Lcd_DisplayString (const u8* str);
-void Lcd_DisplayNumber (s32 num);
+void Lcd_DisplayNumber (s64 num);
 void Lcd_DisplayClear(void);
 void Lcd_SetCursorPosition(u8 row, u8 col);
 void Lcd_DisplayShiftLeft(u8 counter);
