+++ ECLIPSE_PROJECTS/AVR_DRIVERS/inc/Usart.h
@@ -30,6 +30,6 @@ void Usart_SetCallbackReceive(void (*callback) (u8 data));
 void Usart_SetCallbackTransmit(void (*callback) (void));
 void Usart_SetCallbackEmpty(void (*callback) (void));
 void Usart_SendString(const u8 *str);
-void Usart_SendNumber(s32 num);
+void Usart_SendNumber(s64 num);
 void Usart_Print(const u8 *str, ...);
 #endif /* INC_USART_H_ */
