+++ ECLIPSE_PROJECTS/AVR_DRIVERS/src/Lcd.c
@@ -95,9 +95,9 @@ void Lcd_DisplayControl(Lcd_DisplayControlType display)
     }
 }
 
-void Lcd_DisplayNumber(s32 num)
+void Lcd_DisplayNumber(s64 num)
 {
-    u32 reversed = 0;
+    u64 reversed = 0;
     u8 counter = 0;
     if (num < 0)
     {
